Lalit Kumar/aqiproject/requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
requests==2.31.0
# pip install mysql-connector-python


